Delete duplicate file
